+++ backend/app/models/Trophy/model.py
@@ -8,7 +8,7 @@ class Trophy(Base):
     __tablename__ = "trophy"
 
     id = Column(Integer, primary_key=True, index=True)
-    id_league = Column(SmallInteger, ForeignKey("league.id"), unique=True, nullable=False)
+    id_league = Column(Integer, ForeignKey("league.id"), unique=True, nullable=False)
     name = Column(String(255))
     image = Column(String(255))
